Minor tweaks to benchmarks
