+++ benchmark/simple/index.js
@@ -1,0 +1,1 @@
+import './simple.jsx';
