+++ benchmark/static/index.js
@@ -1,0 +1,1 @@
+import './static.jsx';
